# backend/docs/swagger.yaml
basePath: /v1
definitions:
  entities.Author:
    properties:
      icon_url:
        type: string
      login:
        type: string
    type: object
  entities.Comment:
    properties:
      author:
        $ref: '#/definitions/entities.Author'
      created_at:
        type: string
      id:
        type: integer
      text:
        maxLength: 250
        minLength: 1
        type: string
      updated_at:
        type: string
    required:
    - text
    type: object
...
  entities.FullRecipe:
    properties:
      author:
        $ref: '#/definitions/entities.Author'
      comments:
        items:
          $ref: '#/definitions/entities.Comment'
        type: array
      is_liked:
        type: boolean
      likes_count:
        type: integer
      recipe:
        $ref: '#/definitions/entities.Recipe'
    type: object
...
  entities.Recipe:
    properties:
      about:
        maxLength: 2500
        type: string
      complexitiy:
        enum:
        - 1
        - 2
        - 3
        maximum: 3
        minimum: 1
        type: integer
      created_at:
        type: string
      id:
        type: integer
      ingridients:
        maxLength: 1500
        type: string
      need_time:
        type: string
      photos_urls:
        type: string
      title:
        maxLength: 50
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - about
    - complexitiy
    - ingridients
    - need_time
    - title
    type: object
  entities.RecipeFilter:
    properties:
      limit:
        example: 25
        type: integer
      offset:
        example: 0
        type: integer
      order_by:
        enum:
        - -1
        - 0
        - 1
        maximum: 1
        minimum: -1
        type: integer
      order_field:
        enum:
        - title
        - about
        - ingridients
        - emtpy
        example: title
        type: string
      query:
        example: tasty food
        type: string
    type: object
  entities.RecipeInfo:
    properties:
      info:
        $ref: '#/definitions/entities.FullRecipe'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: RestAPI for recipe site
  title: RecipeSite
  version: "1.0"
paths:
...
  /recipe:
    get:
      description: Get all recipe
      operationId: get all recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Recipe'
            type: array
        "500":
          description: Internal Server Error
      summary: Get all recipe
      tags:
      - recipe
    post:
      consumes:
      - application/json
      description: Get filtered recipe
      operationId: get filtered recipe
      parameters:
      - description: filter
        in: body
        name: filter
        schema:
          $ref: '#/definitions/entities.RecipeFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Recipe'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get filtered recipe
      tags:
      - recipe
  /recipe/{id}:
    get:
      description: Get recipe
      operationId: get recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RecipeInfo'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get recipe
      tags:
      - recipe
...
  /user/{login}/recipe:
    post:
      consumes:
      - multipart/form-data
      description: Create recipe
      operationId: create recipe
      parameters:
      - description: Photos
        in: formData
        name: photos
        type: file
      - in: formData
        maxLength: 2500
        name: about
        required: true
        type: string
      - enum:
        - 1
        - 2
        - 3
        in: formData
        maximum: 3
        minimum: 1
        name: complexitiy
        required: true
        type: integer
      - in: formData
        maxLength: 1500
        name: ingridients
        required: true
        type: string
      - in: formData
        name: need_time
        required: true
        type: string
      - in: formData
        maxLength: 50
        minLength: 3
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: Create recipe
      tags:
      - recipe
  /user/{login}/recipe/{id}:
    delete:
      description: Delete recipe
      operationId: delete recipe
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete recipe
      tags:
      - recipe
    put:
      consumes:
      - multipart/form-data
      description: Update recipe
      operationId: update recipe
      parameters:
      - description: Photos
        in: formData
        name: photos
        type: file
      - in: formData
        maxLength: 2500
        name: about
        type: string
      - enum:
        - 1
        - 2
        - 3
        in: formData
        maximum: 3
        minimum: 1
        name: complexitiy
        type: integer
      - in: formData
        maxLength: 1500
        name: ingridients
        type: string
      - in: formData
        name: need_time
        type: string
      - in: formData
        maxLength: 50
        minLength: 3
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update recipe
      tags:
      - recipe
swagger: "2.0"